+++ chalicelib/processors/SummaryProcessor.py
@@ -2,6 +2,7 @@ import math
 
 from chalicelib.processors.GenericProcessor import GenericProcessor, key_ChunkedInputs, key_ChunkPrefix, key_IsChunked, key_NumberOfChunks
 from chalicelib.version import API_VERSION
+from chalicelib.usage import OpenAIDefaults
 
 
 class SummarizeProcessor(GenericProcessor):
@@ -9,7 +10,8 @@ class SummarizeProcessor(GenericProcessor):
         super().__init__(API_VERSION, {
             'main': 'summarize.prompt',
             'role_system': 'summarize-role-system.prompt'
-        })
+        }, {'model': OpenAIDefaults.boost_default_gpt_model,
+            'temperature': OpenAIDefaults.temperature_medium_with_explanation})
 
     def calculate_input_token_buffer(self, total_max) -> int:
         # we'll leave 90% of the buffer for the input, and 10% for the output